+++ types.d.ts
@@ -1,0 +1,3 @@
+declare module '@algolia/autocomplete-theme-classic' {
+  // CSS only
+}