+++ src/client/theme/SearchBar/HighlightSearchResults.tsx
@@ -3,7 +3,6 @@ import { useEffect, useState } from "react";
 import { useLocation } from "@docusaurus/router";
 import useDocusaurusContext from "@docusaurus/useDocusaurusContext";
 import { docsBasePath, blogBasePath } from "./generatedWrapper";
-// @ts-expect-error
 import { useHistory } from "@docusaurus/router";
 
 function isDocsOrBlog(baseUrl: string) {
@@ -14,24 +13,26 @@ function isDocsOrBlog(baseUrl: string) {
 }
 
 export function HighlightSearchResults() {
-  const location = useLocation();
+  const location = useLocation<{ cmfcmfhighlight?: string[] }>();
   const history = useHistory();
   const {
     siteConfig: { baseUrl },
   } = useDocusaurusContext();
-
-  // @ts-expect-error
-  const termsToHighlight = location.state?.cmfcmfhighlight ?? [];
-
   const [terms, setTerms] = useState<string[]>([]);
 
   useEffect(() => {
+    const termsToHighlight = location.state?.cmfcmfhighlight ?? [];
     if (termsToHighlight.length === 0) {
       return;
     }
     setTerms(termsToHighlight);
-    history.replace(location.href);
-  }, [termsToHighlight, history, location]);
+
+    const { cmfcmfhighlight, ...state } = location.state;
+    history.replace({
+      ...location,
+      state,
+    });
+  }, [location.state?.cmfcmfhighlight, history, location]);
 
   useEffect(() => {
     if (terms.length === 0) {
@@ -50,7 +51,7 @@ export function HighlightSearchResults() {
     const options = {
       ignoreJoiners: true,
     };
-    mark.mark(termsToHighlight, options);
+    mark.mark(terms, options);
     return () => mark.unmark(options);
   }, [terms]);